+++ src/logic/LogicBuilder.ts
@@ -3,7 +3,7 @@ import { CityName } from "./CityName";
 import { IMainSceneParams } from "./IMainSceneParams";
 import { Logic } from "./Logic";
 import { Player } from "./Player";
-import { Ware } from "./ware";
+import { Ware } from "./Ware";
 
 export class LogicBuilder {
     public static create(): IMainSceneParams {
