+++ src/logic/Ware.config.test.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { Ware } from "./ware";
+import { Ware } from "./Ware";
 import { WareType } from "./WareType";
 
 describe("Ware config - every WareType has set a value", () => {
